+++ src/briefy/ws/utils/filter.py
@@ -29,6 +29,7 @@ class COMPARISON(Enum):
     EQ = 'eq'
     GT = 'gt'
     IN = 'in_'
+    LIKE = 'like'
     EXCLUDE = 'notin_'
 
 
@@ -65,7 +66,7 @@ def create_filter_from_query_params(query_params: dict, allowed_fields: list):
             )
             continue
 
-        m = re.match(r'^(min|max|not|lt|gt|in|exclude)_(\w+)$', param)
+        m = re.match(r'^(min|max|not|lt|gt|in|exclude|like)_(\w+)$', param)
         if m:
             keyword, field = m.groups()
             operator = getattr(COMPARISON, keyword.upper())
@@ -82,6 +83,8 @@ def create_filter_from_query_params(query_params: dict, allowed_fields: list):
         value = data.native_value(param_value, field)
         if operator in (COMPARISON.IN, COMPARISON.EXCLUDE):
             value = set([data.native_value(v, field) for v in param_value.split(',')])
+        elif operator in (COMPARISON.LIKE, ):
+            value = '%{value}%'.format(value=value.replace('%', ''))
         filters.append(Filter(field, value, operator))
     return filters